+++ timeloop-basic.c
@@ -30,8 +30,8 @@ my_pipe (int *fds)
 int
 read_all (int fd, char *buf, int len)
 {
-  int bytes_read = 0;
-  int count;
+  size_t bytes_read = 0;
+  ssize_t count;
 
   while (bytes_read < len)
     {
@@ -53,8 +53,8 @@ read_all (int fd, char *buf, int len)
 int
 write_all (int fd, char *buf, int len)
 {
-  int bytes_written = 0;
-  int count;
+  size_t bytes_written = 0;
+  ssize_t count;
 
   while (bytes_written < len)
     {
